Issue #4746: Move CrashReporter initialization to Components class.
