+++ app/src/main/java/org/mozilla/focus/Components.kt
@@ -4,7 +4,9 @@
 
 package org.mozilla.focus
 
+import android.app.PendingIntent
 import android.content.Context
+import android.content.Intent
 import mozilla.components.browser.session.Session
 import mozilla.components.browser.session.SessionManager
 import mozilla.components.browser.session.engine.EngineMiddleware
@@ -27,12 +29,18 @@ import mozilla.components.feature.session.SettingsUseCases
 import mozilla.components.feature.session.TrackingProtectionUseCases
 import mozilla.components.feature.tabs.CustomTabsUseCases
 import mozilla.components.feature.tabs.TabsUseCases
+import mozilla.components.lib.crash.CrashReporter
+import mozilla.components.lib.crash.service.CrashReporterService
+import mozilla.components.lib.crash.service.MozillaSocorroService
+import mozilla.components.lib.crash.service.SentryService
 import mozilla.components.service.location.LocationService
+import org.mozilla.focus.activity.MainActivity
 import org.mozilla.focus.components.EngineProvider
 import org.mozilla.focus.downloads.DownloadService
 import org.mozilla.focus.engine.ClientWrapper
 import org.mozilla.focus.engine.LocalizedContentInterceptor
 import org.mozilla.focus.engine.SanityCheckMiddleware
+import org.mozilla.focus.locale.LocaleManager
 import org.mozilla.focus.notification.PrivateNotificationMiddleware
 import org.mozilla.focus.search.SearchFilterMiddleware
 import org.mozilla.focus.search.SearchMigration
@@ -41,6 +49,7 @@ import org.mozilla.focus.state.AppStore
 import org.mozilla.focus.state.Screen
 import org.mozilla.focus.telemetry.TelemetryMiddleware
 import org.mozilla.focus.utils.Settings
+import java.util.Locale
 
 /**
  * Helper object for lazily initializing components.
@@ -129,6 +138,8 @@ class Components(
     @Suppress("DEPRECATION")
     val customTabsUseCases: CustomTabsUseCases by lazy { CustomTabsUseCases(sessionManager, sessionUseCases.loadUrl) }
 
+    val crashReporter: CrashReporter by lazy { createCrashReporter(context) }
+
     @Deprecated("Use BrowserStore instead")
     private val sessionManager by lazy {
         SessionManager(engine, store)
@@ -140,5 +151,65 @@ private fun determineInitialScreen(context: Context): Screen {
         Screen.FirstRun
     } else {
         Screen.Home
+    }
+}
+
+private fun createCrashReporter(context: Context): CrashReporter {
+    val services = mutableListOf<CrashReporterService>()
+
+    if (!BuildConfig.SENTRY_TOKEN.isNullOrEmpty()) {
+        val sentryService = SentryService(
+            context,
+            BuildConfig.SENTRY_TOKEN,
+            tags = mapOf(
+                "build_flavor" to BuildConfig.FLAVOR,
+                "build_type" to BuildConfig.BUILD_TYPE,
+                "locale_lang_tag" to getLocaleTag(context)
+            ),
+            environment = BuildConfig.BUILD_TYPE,
+            sendEventForNativeCrashes = false // Do not send native crashes to Sentry
+        )
+
+        services.add(sentryService)
+    }
+
+    val socorroService = MozillaSocorroService(
+        context,
+        appName = "Focus",
+        version = org.mozilla.geckoview.BuildConfig.MOZ_APP_VERSION,
+        buildId = org.mozilla.geckoview.BuildConfig.MOZ_APP_BUILDID,
+        vendor = org.mozilla.geckoview.BuildConfig.MOZ_APP_VENDOR,
+        releaseChannel = org.mozilla.geckoview.BuildConfig.MOZ_UPDATE_CHANNEL
+    )
+    services.add(socorroService)
+
+    val intent = Intent(context, MainActivity::class.java).apply {
+        flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TOP
+    }
+
+    val pendingIntent = PendingIntent.getActivity(
+        context,
+        0,
+        intent,
+        0)
+
+    return CrashReporter(
+        context = context,
+        services = services,
+        promptConfiguration = CrashReporter.PromptConfiguration(
+            appName = context.resources.getString(R.string.app_name)
+        ),
+        shouldPrompt = CrashReporter.Prompt.ALWAYS,
+        enabled = true,
+        nonFatalCrashIntent = pendingIntent
+    )
+}
+
+private fun getLocaleTag(context: Context): String {
+    val currentLocale = LocaleManager.getInstance().getCurrentLocale(context)
+    return if (currentLocale != null) {
+        currentLocale.toLanguageTag()
+    } else {
+        Locale.getDefault().toLanguageTag()
     }
 }
